Rename this class.
